pyproject.toml
[tool.poetry]
name = "doctoraizer"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8.5"
Pillow = "^8.2.0"
python-docx = "^0.8.11"
numpy = "1.19.5"
tensorflow = "^2.5.0"
pycrypto = "^2.6.1"
python-decouple = "^3.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"